+++ app/api/exams/by-user/[userId]/route.ts
@@ -1,4 +1,4 @@
-import { examService } from "@/services/examService"
+import { examService } from "@/services/api/examService"
 import { ApiResponse, ApiResponseError } from "@/types/api"
 import { Exam } from "@/types/models"
 import { NextRequest, NextResponse } from "next/server"
